Refactor alarms module to domain-driven architecture and configure TypeScript watch options
- Move AlarmsModule import path to application layer structure
- Add Alarm domain entity with id, name, and severity properties
- Configure TypeScript exclude patterns for node_modules and dist
- Add watchOptions with dynamic priority polling for improved development experience
- Remove test spec files and entity boilerplate from alarms module
