--- advenced-architecture/src/alarms/alarms.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { AlarmsService } from './alarms.service';
-
-describe('AlarmsService', () => {
-  let service: AlarmsService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [AlarmsService],
-    }).compile();
-
-    service = module.get<AlarmsService>(AlarmsService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
