--- advenced-architecture/src/alarms/entities/alarm.entity.ts
@@ -1,1 +1,0 @@
-export class Alarm {}
