+++ advenced-architecture/src/app.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { AlarmsModule } from './alarms/alarms.module';
+import { AlarmsModule } from './alarms/application/alarms.module';
 
 @Module({
   imports: [AlarmsModule],
